\name{add_character_data}
\alias{add_character_data}
\title{Add character data to a nexml object}
\usage{
add_character_data(x, nexml = new("nexml"), ...)
}
\arguments{
  \item{x}{character data, in which character traits labels
  are column names and taxon labels are row names.  x can
  be in matrix or data.frame format.}
}
\description{
  Add character data to a nexml object
}

