% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_trees}
\alias{add_trees}
\title{add_trees}
\usage{
add_trees(phy, nexml = new("nexml"), append_to_existing_otus = FALSE)
}
\arguments{
\item{phy}{a phylo object, multiPhylo object, or list of
mulitPhylo to be added to the nexml}

\item{nexml}{a nexml object to which we should append this phylo.
By default, a new nexml object will be created.}

\item{append_to_existing_otus}{logical, indicating if we should
make a new OTU block (default) or append to the existing one.}
}
\value{
a nexml object containing the phy in nexml format.
}
\description{
add_trees
}
\examples{
library("geiger")
data(geospiza)
geiger_nex <- add_trees(geospiza$phy)
}

