% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Annotated-class}
\alias{Annotated-class}
\title{Class of objects that have metadata as lists of meta elements}
\description{
Class of objects that have metadata as lists of meta elements
}
\section{Slots}{

\describe{
\item{\code{meta}}{list of \code{meta} objects}

\item{\code{about}}{for RDF extraction, the identifier for the resource that this
object is about}
}}

