% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coalesce_}
\alias{coalesce_}
\title{Front-end to dplyr::coalesce to deal with NULL vectors}
\usage{
coalesce_(...)
}
\arguments{
\item{...}{the vectors to coalesce on NA}
}
\value{
a vector of the same type and length as the last argument
}
\description{
Replaces any NULL argument with a vector of \code{NA}, and casts every vector
to the same type as the last vector. After that, calls \code{\link[dplyr:coalesce]{dplyr::coalesce()}}.
}
\seealso{
\code{\link[dplyr:coalesce]{dplyr::coalesce()}}
}
