% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R, R/meta.R
\name{nexml.tree}
\alias{nexml.tree}
\alias{nexml.trees}
\alias{nexml.node}
\alias{nexml.edge}
\alias{nexml.otu}
\alias{nexml.otus}
\alias{nexml.char}
\alias{nexml.characters}
\alias{nexml.format}
\alias{nexml.state}
\alias{nexml.uncertain_state}
\alias{nexml.states}
\alias{nexml.uncertain_states}
\alias{nexml.polymorphic_states}
\alias{nexml.member}
\alias{nexml.matrix}
\alias{nexml.row}
\alias{nexml.seq}
\alias{nexml.cell}
\alias{nexml.meta_}
\title{Constructor for the respective class}
\usage{
nexml.tree(...)

nexml.trees(...)

nexml.node(...)

nexml.edge(...)

nexml.otu(...)

nexml.otus(...)

nexml.char(...)

nexml.characters(...)

nexml.format(...)

nexml.state(...)

nexml.uncertain_state(...)

nexml.states(...)

nexml.uncertain_states(...)

nexml.polymorphic_states(...)

nexml.member(...)

nexml.matrix(...)

nexml.row(...)

nexml.seq(...)

nexml.cell(...)
}
\arguments{
\item{...}{optionally, parameters passed on to \code{\link[methods:new]{new()}}}
}
\description{
Creates an instance of the class corresponding to the respective NeXML
element, and initializes its slots with the provided parameters, if any.
}
\details{
Usually, users won't need to invoke this directly.
}
\seealso{
\link[=meta]{nexml.meta()} for documentation of \code{nexml.meta()}
}
