% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-utils.R
\name{.sigLabel}
\alias{.sigLabel}
\title{Create a label for a method signature}
\usage{
.sigLabel(signature)
}
\arguments{
\item{signature}{the signature for which to create a label, as a vector
or list of strings, or as an instance of \link[methods:signature-class]{signature}.}
}
\value{
a character string
}
\description{
Creates a label for a signature mirroring the result of \code{.sigLabel()}
in the \code{methods} package, which unfortunately does not export the function.
This is needed, for example, for the \code{excluded} slot in the
\link[methods:MethodWithNext-class]{MethodWithNext} class.
}
