\name{edge-class}
\alias{edge-class}
\title{Strucrure of meta S4 class.}
\description{
  Class will be hold information about metadata of some
  node parsed from NeXML tree in DOM mode.
}
\details{
  \describe{ \item{source}{Value of edge class}

  \item{target}{Target}

  \item{id}{An identifying number of edge instance}

  \item{length}{Edge lenght}

  }
}

