% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixed-uris.R
\name{expand_prefix}
\alias{expand_prefix}
\title{Expand namespace-prefixed string}
\usage{
expand_prefix(x, namespaces = NULL)
}
\arguments{
\item{x}{a character vector of potentially namespace-prefixed strings}

\item{namespaces}{a named vector of namespaces, with namespace prefixes
being the names. A "base" namespace with an empty name can be included.
If not provided, or if empty, the input vector is returned as is.}
}
\value{
a character vector, of the same length as the input vector
}
\description{
Substitutes the namespace prefix in the input vector of strings with
the corresponding namespaces.
}
\details{
Namespace prefixes are expected to be separated by one or more semicolons.
Prefixes that cannot be matched to the vector of namespaces will be left
as is. For strings that do not have a namespace prefix, the vector of
namespaces can contain a base namespace, identified as not having a name,
with which these strings will be expanded.
}
