% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_all_meta}
\alias{get_all_meta}
\title{Get flattened list of meta annotations}
\usage{
get_all_meta(annotated)
}
\arguments{
\item{annotated}{the object from which to extract meta object annotations}
}
\value{
a flat list of \code{meta} objects
}
\description{
Collects recursively (in the case of nested meta annotations) all meta
object annotations for the given object, and returns the result as a flat
list.
}
\details{
Does not check that the input object can actually have meta annotations.
An invalid slot error will be generated if it can't.
}
