\name{get_characters_list}
\alias{get_characters_list}
\title{Extract the character matrix}
\usage{
get_characters_list(nexml)
}
\arguments{
  \item{nexml}{a nexml object}

  \item{nexml}{nexml object (e.g. from read.nexml)}
}
\value{
the list of taxa
}
\description{
Extract the character matrix
}

