% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_flat_trees}
\alias{get_flat_trees}
\title{get_flat_trees}
\usage{
get_flat_trees(nexml)
}
\arguments{
\item{nexml}{a representation of the nexml object from  which the data is to be retrieved}
}
\value{
a multiPhylo object (list of ape::phylo objects).  See details.
}
\description{
extract a single multiPhylo object containing all trees in the nexml
}
\details{
Note that this method collapses any heirachical structure that may have been present as multiple `trees` nodes in the original nexml (though such a feature is rarely used).  To preserve that structure, use \code{\link{get_trees}} instead.
}
\seealso{
\code{\link{get_trees}} \code{\link{get_trees}} \code{\link{get_item}}
}

