\name{get_item}
\alias{get_item}
\title{Get the desired element from the nexml object}
\usage{
  get_item(nexml,
    element = c("tree", "trees", "metadata", "otu"),
    level = c("nexml", "trees", "tree", "otus", "otu"))
}
\arguments{
  \item{element}{}
}
\description{
  Get the desired element from the nexml object
}

