\docType{methods}
\name{get_metadata,nexml-method}
\alias{get_metadata,nexml-method}
\title{get all top-level metadata  More extensible than hardwired functions}
\usage{
\S4method{get_metadata}{nexml}(object, level = c("nexml", "otus", "otu",
  "trees", "tree", "edge", "node"))
}
\description{
  get all top-level metadata More extensible than hardwired
  functions
}

