\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(nexml, level = c("nexml", "otus", "otu", "trees", "tree",
  "characters", "format", "states", "all"))
}
\arguments{
  \item{nexml}{a nexml object}

  \item{level}{the name of the level of element desired.
  all corresponds to all meta elements.}
}
\value{
the requested metadata
}
\description{
get_metadata
}

