% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_namespaces.R
\name{get_namespaces}
\alias{get_namespaces}
\title{get namespaces}
\usage{
get_namespaces(nexml)
}
\arguments{
\item{nexml}{a nexml object}
}
\value{
a named character vector providing the URLs defining each
of the namespaces used in the nexml file.  Names correspond to
the prefix abbreviations of the namespaces.
}
\description{
get namespaces
}
\examples{
comp_analysis <- system.file("examples", "comp_analysis.xml", package="RNeXML")
nex <- nexml_read(comp_analysis)
get_namespaces(nex)
}

