\name{get_otu}
\alias{get_otu}
\alias{get_taxa}
\title{Retrieve names of all species/otus otus (operational taxonomic units) included in the nexml}
\usage{
get_otu(object)
}
\description{
  Retrieve names of all species/otus otus (operational
  taxonomic units) included in the nexml
}
\seealso{
  \code{\link{get_item}}
}

