\name{get_otu_maps}
\alias{get_otu_maps}
\title{get otus map}
\usage{
get_otu_maps(nexml)
}
\arguments{
  \item{nexml}{nexml object}
}
\value{
  a list showing the mapping between (internal) otu
  identifiers and labels (taxonomic names). List is named
  by the id of the otus block.
}
\description{
  get otus map
}
\details{
  largely for internal use
}

