\name{get_otus_list}
\alias{get_otus_list}
\alias{get_taxa_list}
\title{Retrieve names of all species/otus otus (operational taxonomic units) included in the nexml}
\usage{
get_otus_list(object)
}
\description{
  Retrieve names of all species/otus otus (operational
  taxonomic units) included in the nexml
}
\seealso{
  \code{\link{get_item}}
}

