\name{get_taxa}
\alias{get_otu}
\alias{get_taxa}
\title{Retrieve names of all species/otus otus (operational taxonomic units) included in the nexml}
\usage{
get_taxa(nexml)
}
\description{
Retrieve names of all species/otus otus (operational
taxonomic units) included in the nexml
}
\seealso{
\code{\link{get_item}}
}

