% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_taxa_meta_list}
\alias{get_taxa_meta_list}
\title{get_taxa_meta_list}
\usage{
get_taxa_meta_list(nexml, what = "href")
}
\arguments{
\item{nexml}{a nexml object}

\item{what}{One of href, rel, id, or xsi:type}
}
\value{
the list of metadata for each taxon
}
\description{
Retrieve metadata of all species/otus otus (operational taxonomic units) included in the nexml
}
\examples{
\dontrun{
data(bird.orders)
birds <- add_trees(bird.orders)
birds <- taxize_nexml(birds, "NCBI")
RNeXML:::get_taxa_meta_list(birds)
RNeXML:::get_taxa_meta_list(birds, 'rel')
RNeXML:::get_taxa_meta_list(birds, 'id')
RNeXML:::get_taxa_meta_list(birds, 'xsi:type')
}
}
\seealso{
\code{\link{get_item}}
}
\keyword{internal}

