\name{meta-class}
\alias{meta-class}
\title{Strucrure of meta S4 class.}
\description{
  Class will be hold information about metadata of some
  node parsed from NeXML tree in DOM mode.
}
\details{
  \describe{ \item{id}{An identifyong number of meta
  instance.}

  \item{property}{A propetry of metadata.}

  \item{content}{Content of metadata class.}

  \item{datatype}{Type of presented metadata.}

  }
}

