% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmap.R
\name{nexml_to_simmap}
\alias{nexml_to_simmap}
\title{Convert nexml object with simmap to phylo}
\usage{
nexml_to_simmap(nexml)
}
\arguments{
\item{nexml}{a nexml object}
}
\value{
a simmap object (phylo object with a \code{$maps} element
for use in phytools functions).
}
\description{
Convert nexml object with simmap to phylo
}
