% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nexml_validate}
\alias{nexml_validate}
\title{validate nexml using the online validator tool}
\usage{
nexml_validate(file)
}
\arguments{
\item{file}{path to the nexml file to validate}
}
\value{
TRUE if the file is valid, false otherwise
}
\description{
validate nexml using the online validator tool
}
\details{
Requires an internet connection.  see http://www.nexml.org/nexml/phylows/validator for more information in debugging invalid files
}

