\name{node-class}
\alias{node-class}
\title{Strucrure of node S4 class.}
\description{
  Class will be hold information about node from NeXML tree
  in DOM mode.
}
\details{
  \describe{ \item{id}{An identifying number of node.}

  \item{label}{Name of the node (tree, otu, etc).}

  \item{otu}{Value of the 'otu' attribute}
  \item{about}{Node description}

  \item{meta}{Metadata that is presented by meta class.} }
}

