% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmap.R
\name{simmap_to_nexml}
\alias{simmap_to_nexml}
\title{Convert phylo with attached simmap to nexml object}
\usage{
simmap_to_nexml(phy, state_ids = NULL)
}
\arguments{
\item{phy}{a phylo object containing simmap `phy$maps`` element,
from the phytools package}

\item{state_ids}{a named character vector giving the state
names corresponding to the ids used to refer to each state
in nexml.  If null ids will be generated and states taken from
the phy$states names.}
}
\value{
a nexml representation of the simmap
}
\description{
Convert phylo with attached simmap to nexml object
}
\examples{
simmap_ex <- read.nexml(system.file("examples","simmap_ex.xml", package="RNeXML"))
phy <- nexml_to_simmap(simmap_ex)
nex <- simmap_to_nexml(phy) 
}
