\name{toPhylo}
\alias{toPhylo}
\title{nexml to phylo}
\usage{
toPhylo(tree, otus)
}
\arguments{
  \item{tree}{an nexml tree element (nexml}

  \item{otus}{a character string of taxonomic labels, named
  by the otu ids. e.g. (from get_otu_maps for the otus set
  matching the relevant trees node.}
}
\value{
phylo object
}
\description{
nexml to phylo coercion
}

