\name{tree-class}
\alias{tree-class}
\title{Strucrure of tree S4 class.}
\description{
  Class will be hold information about tree parsed from
  NeXML file in DOM mode.
}
\details{
  \describe{ \item{id}{An identification value of tree}

  \item{xsi:type}{Datatype value}

  \item{label}{Label value}

  \item{nodes}{Set of node items}

  \item{edges}{Set of edge items}

  }
}

