% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{methods}
\name{c,meta-method}
\alias{c,meta-method}
\title{Concatenate meta elements into a ListOfmeta}
\usage{
\S4method{c}{meta}(x, ..., recursive = FALSE)
}
\arguments{
\item{x, ...}{meta elements to be concatenated, e.g. see \code{\link{meta}}}

\item{recursive}{logical, if 'recursive=TRUE', the function
descends through lists and combines their elements into a vector.}
}
\value{
a listOfmeta object containing multiple meta elements.
}
\description{
Concatenate meta elements into a ListOfmeta
}
\examples{
c(meta(content="example", property="dc:title"),
  meta(content="Carl", property="dc:creator"))

}
