% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trees.R
\name{flatten_multiphylo}
\alias{flatten_multiphylo}
\title{Flatten a multiphylo object}
\usage{
flatten_multiphylo(object)
}
\arguments{
\item{object}{a list of multiphylo objects}
}
\description{
Flatten a multiphylo object
}
\details{
NeXML has the concept of multiple <trees> nodes, each with multiple child <tree> nodes.
This maps naturally to a list of multiphylo  objects.  Sometimes
this hierarchy conveys important structural information, so it is not discarded by default. 
Occasionally it is useful to flatten the structure though, hence this function.  Note that this
discards the original structure, and the nexml file must be parsed again to recover it.
}
