% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characters.R
\name{get_characters}
\alias{get_characters}
\title{Get character data.frame from nexml}
\usage{
get_characters(nex, rownames_as_col = FALSE, otu_id = FALSE,
  otus_id = FALSE, include_state_types = FALSE)
}
\arguments{
\item{nex}{a nexml object}

\item{rownames_as_col}{option to return character matrix rownames (with taxon ids) as it's own column in the 
data.frame. Default is FALSE for compatibility with geiger and similar packages.}

\item{otu_id}{logical, default FALSE. return a column with the 
otu id (for joining with otu metadata, etc)}

\item{otus_id}{logical, default FALSE. return a column with the 
otus block id (for joining with otu metadata, etc)}

\item{include_state_types}{logical, default FALSE. whether to also return a 
matrix of state types (with values standard, polymorphic, and uncertain)}
}
\value{
the character matrix as a data.frame, or if `include_state_types is
 TRUE a list of two elements, `characters` as the character matrix, and
 state_types as a matrix of state types. Both matrices will be in the same
 ordering of rows and columns.
}
\description{
Get character data.frame from nexml
}
\details{
RNeXML will attempt to return the matrix using the NeXML taxon (otu) labels to name the rows
 and the NeXML char labels to name the traits (columns).  If these are unavailable or not unique, the NeXML
 id values for the otus or traits will be used instead.
}
\examples{
\dontrun{
# A simple example with a discrete and a continous trait
f <- system.file("examples", "comp_analysis.xml", package="RNeXML")
nex <- read.nexml(f)
get_characters(nex)

# A more complex example -- currently ignores sequence-type characters
f <- system.file("examples", "characters.xml", package="RNeXML")
nex <- read.nexml(f)
get_characters(nex)
}
}
