% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_characters_list}
\alias{get_characters_list}
\title{Extract the character matrix}
\usage{
get_characters_list(nexml)
}
\arguments{
\item{nexml}{nexml object (e.g. from read.nexml)}
}
\value{
the list of taxa
}
\description{
Extract the character matrix
}
\examples{
comp_analysis <- system.file("examples", "comp_analysis.xml", package="RNeXML")
nex <- nexml_read(comp_analysis)
get_characters_list(nex)
}

