% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_characters_list}
\alias{get_characters_list}
\title{Extract the character matrix}
\usage{
get_characters_list(nexml, rownames_as_col = FALSE)
}
\arguments{
\item{nexml}{nexml object (e.g. from read.nexml)}

\item{rownames_as_col}{option to return character matrix rownames
(with taxon ids) as it's own column in the data.frame. Default is FALSE
for compatibility with geiger and similar packages.}
}
\value{
the list of taxa
}
\description{
Extract the character matrix
}
\examples{
comp_analysis <- system.file("examples", "comp_analysis.xml", package="RNeXML")
nex <- nexml_read(comp_analysis)
get_characters_list(nex)
}
