% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(nexml, level = "nexml")
}
\arguments{
\item{nexml}{a nexml object}

\item{level}{the name of the level of element desired, or
use "all" to access all meta elements.}
}
\value{
the requested metadata
}
\description{
get_metadata
}
\examples{
comp_analysis <- system.file("examples", "comp_analysis.xml", package="RNeXML")
nex <- nexml_read(comp_analysis)
get_metadata(nex)
}

