% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_taxa}
\alias{get_otu}
\alias{get_taxa}
\title{get_taxa}
\usage{
get_taxa(nexml)
}
\arguments{
\item{nexml}{a nexml object}
}
\value{
the list of taxa
}
\description{
Retrieve names of all species/otus otus (operational taxonomic units) included in the nexml
}
\examples{
comp_analysis <- system.file("examples", "comp_analysis.xml", package="RNeXML")
nex <- nexml_read(comp_analysis)
get_taxa(nex)
}
\seealso{
\code{\link{get_item}}
}

