% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_taxa.R
\name{get_taxa_list}
\alias{get_otus_list}
\alias{get_taxa_list}
\title{get_taxa_list}
\usage{
get_taxa_list(nexml)
}
\arguments{
\item{nexml}{a nexml object}
}
\value{
the list of taxa
}
\description{
Retrieve names of all species/otus otus (operational taxonomic units) included in the nexml
}
\seealso{
\code{\link{get_item}}
}

