% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_taxa_meta.R
\name{get_taxa_meta}
\alias{get_taxa_meta}
\title{get_taxa_meta}
\usage{
get_taxa_meta(nexml, what = "href")
}
\arguments{
\item{nexml}{a nexml object}

\item{what}{One of href, rel, id, or xsi:type}
}
\value{
the list of metadata for each taxon
}
\description{
Retrieve metadata of all species/otus otus (operational taxonomic units) included in the nexml
}
\examples{
\dontrun{
data(bird.orders)
birds <- add_trees(bird.orders)
birds <- taxize_nexml(birds, "NCBI")
RNeXML:::get_taxa_meta(birds)
RNeXML:::get_taxa_meta(birds, 'rel')
RNeXML:::get_taxa_meta(birds, 'id')
RNeXML:::get_taxa_meta(birds, 'xsi:type')
 }
}
\seealso{
\code{\link{get_item}}
}
\keyword{internal}

