% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nexml_to_simmap}
\alias{nexml_to_simmap}
\title{nexml_to_simmap}
\usage{
nexml_to_simmap(nexml)
}
\arguments{
\item{nexml}{a nexml object}
}
\value{
a simmap object (phylo object with a $maps element
 for use in phytools functions).
}
\description{
nexml_to_simmap
}
\examples{
data(simmap_ex)
phy <- nexml_to_simmap(simmap_ex)
nex <- simmap_to_nexml(phy)
}

