% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_trees.R
\name{toPhylo}
\alias{toPhylo}
\title{nexml to phylo}
\usage{
toPhylo(tree, otus)
}
\arguments{
\item{tree}{an nexml tree element}

\item{otus}{a character string of taxonomic labels, named by the otu ids.
e.g. (from get_otu_maps for the otus set matching the relevant trees node.}
}
\value{
phylo object.  If a "reconstructions" annotation is found on the
edges, return simmap maps slot as well.
}
\description{
nexml to phylo coercion
}

